SafetyProductionSystem/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.0.5
django-appconf==1.0.2
django-bower==5.2.0
django-compressor==2.2
django-crispy-forms==1.7.2
django-el-pagination==3.2.4
django-impersonate==1.3
django-lb-adminlte==0.9.6
django-lbattachment==1.0.4
django-lbutils==1.0.4
django-stronghold==0.3.0
graphviz==0.9
idna==2.7
Jinja2==2.10
jsonfield==2.0.2
MarkupSafe==1.0
Pillow==5.2.0
pycparser==2.18
pygraphviz==1.3.1
PyMySQL==0.9.2
pytz==2018.4
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
XlsxWriter==1.0.7
xlrd==1.1.0
git+https://github.com/18137131021/django-lb-workflow.git@master#egg=django-lb-workflow
git+https://github.com/chengdh/django-crontab.git@chang_run_job#egg=django-crontab
